chore: address PR feedback, organize block components
