+++ nextjs-app/app/components/candidatesBanner/CandidatesBanner.module.css
