+++ nextjs-app/app/components/valueProp/ValuePropImageColumn.tsx
