+++ nextjs-app/app/components/valueProp/ValuePropTestimonialColumn.tsx
