+++ nextjs-app/app/components/valueProp/ValueProposition.tsx
